+++ src/database/factories/user.factory.ts
@@ -1,6 +1,6 @@
+import { ifs } from '@/config/data-sources';
 import { User } from '@/modules/core/entities/user';
 import Chance from 'chance';
-import { getRepository } from 'typeorm';
 
 const chance = new Chance();
 
@@ -10,15 +10,15 @@ export const userFactory = {
       email: chance.email({ domain: 'example.com' }),
       ...attrs
     };
-    return getRepository(User).create(userAttrs);
+    return ifs.getRepository(User).create(userAttrs);
   },
 
   create: async (attrs: Partial<User> = {}): Promise<User> => {
     const user = userFactory.build(attrs);
-    const createdUser = await getRepository(User).save(user);
+    const createdUser = await ifs.getRepository(User).save(user);
     return createdUser;
   },
 
   deleteAll: async (): Promise<void> =>
-    await getRepository(User).query('TRUNCATE "USER" CASCADE')
+    await ifs.getRepository(User).query('TRUNCATE "USER" CASCADE')
 };