+++ src/modules/manufacturing/apm-machine-work-center.rv.ts
@@ -1,8 +1,11 @@
+import { ifs } from '@/config/data-sources';
 import { isAuth } from '@/middlewares/is-auth';
 import { mapError } from '@/utils/map-error';
 import { Arg, Query, Resolver, UseMiddleware } from 'type-graphql';
-import { getConnection } from 'typeorm';
+import { In, Like } from 'typeorm';
+import { MachineMaintenanceView } from './entities/apm-machine-maintenance.vw';
 import { MachineWorkCenter } from './entities/apm-machine-work-center.vt';
+import { IfsWorkCenterView } from './entities/ifs-work-center.vw';
 import { WorkCenterView } from './entities/work-center.vw';
 
 @Resolver()
@@ -14,90 +17,69 @@ export class MachineWorkCenterResolver {
     @Arg('departmentNo') departmentNo: string
   ): Promise<WorkCenterView[] | undefined> {
     try {
-      const sql = `
-        SELECT   work_center_no  AS "workCenterNo",
-                 contract        AS "contract",
-                 description     AS "description",
-                 department_no   AS "departmentNo",
-                 objid           AS "objId"
-        FROM     atj_work_center_v
-        WHERE    contract = :contract
-        AND      department_no = :departmentNo
-        AND      (contract, work_center_no) NOT IN (
-                   SELECT   contract,
-                            work_center_no AS "workCenterNo"
-                   FROM     rob_apm_unavailable_wc_v
-                   WHERE    contract = :contract
-                 )
-      `;
-      const results = await getConnection().query(sql, [
-        contract,
-        departmentNo
-      ]);
-      return results;
+      let sql = '';
+      if (contract === 'AGT') {
+        sql = `
+          SELECT   work_center_no  AS "workCenterNo",
+                   contract        AS "contract",
+                   description     AS "description",
+                   department_no   AS "departmentNo",
+                   objid           AS "objId"
+          FROM     work_center@ifs8agt
+          WHERE    contract = :contract
+          AND      department_no = :departmentNo
+          AND      (contract, work_center_no) NOT IN (
+                    SELECT   contract,
+                             work_center_no AS "workCenterNo"
+                    FROM     rob_apm_unavailable_wc_v
+                    WHERE    contract = :contract
+                   )
+        `;
+      } else {
+        sql = `
+          SELECT   work_center_no  AS "workCenterNo",
+                   contract        AS "contract",
+                   description     AS "description",
+                   department_no   AS "departmentNo",
+                   objid           AS "objId"
+          FROM     work_center
+          WHERE    contract = :contract
+          AND      department_no = :departmentNo
+          AND      (contract, work_center_no) NOT IN (
+                    SELECT   contract,
+                             work_center_no AS "workCenterNo"
+                    FROM     rob_apm_unavailable_wc_v
+                    WHERE    contract = :contract
+                   )
+        `;
+      }
+      const result = await ifs.query(sql, [contract, departmentNo]);
+      return result;
     } catch (err) {
       throw new Error(mapError(err));
     }
   }
 
-  @Query(() => [MachineWorkCenter])
+  @Query(() => [IfsWorkCenterView])
   @UseMiddleware(isAuth)
   async getAssignedMachWcByContractDept(
     @Arg('contract', () => [String]) contract: string[],
-    @Arg('departmentNo', { nullable: true }) departmentNo: string
-  ): Promise<MachineWorkCenter[] | undefined> {
-    let sql = '';
-    let formattedContract = '';
+    @Arg('departmentNo', { defaultValue: '%', nullable: true })
+    departmentNo: string
+  ): Promise<IfsWorkCenterView[] | undefined> {
     try {
-      if (contract[0] === 'AGT') {
-        sql = `
-          SELECT   wc.work_center_no AS "workCenterNo",
-                   wc.contract       AS "contract",
-                   wc.description    AS "description",
-                   wc.department_no  AS "departmentNo",
-                   ram.machine_id    AS "machineId",
-                   ram.description   AS "machineDescription",
-                   wc.objid          AS "objId"
-          FROM     work_center@ifs8agt  wc,
-                   rob_apm_machine      ram
-          WHERE    ram.contract = wc.contract
-          AND      ram.work_center_no = wc.work_center_no
-          AND      wc.department_no LIKE :departmentNo
-        `;
-      } else {
-        sql = `
-          SELECT   wc.work_center_no AS "workCenterNo",
-                   wc.contract       AS "contract",
-                   wc.description    AS "description",
-                   wc.department_no  AS "departmentNo",
-                   ram.machine_id    AS "machineId",
-                   ram.description   AS "machineDescription",
-                   wc.objid          AS "objId"
-          FROM     work_center       wc,
-                   rob_apm_machine   ram
-          WHERE    ram.contract = wc.contract
-          AND      ram.work_center_no = wc.work_center_no
-          AND      wc.department_no LIKE :departmentNo
-        `;
-      }
-      contract.map((item, index) => {
-        formattedContract = `${formattedContract}${
-          index > 0 ? ',' : ''
-        } '${item}'`;
+      const results = await WorkCenterView.find({
+        relations: { machines: true },
+        select: {
+          workCenterNo: true,
+          contract: true,
+          description: true
+        },
+        where: { contract: In(contract), departmentNo: Like(departmentNo) },
+        order: { workCenterNo: 'ASC' }
       });
-      formattedContract.slice(0, formattedContract.length - 1);
-      sql = `${sql}  AND      wc.contract IN (${formattedContract} )`;
-      let results = await getConnection().query(sql, [
-        departmentNo ? departmentNo : '%'
-      ]);
-      if (results) {
-        results = results
-          .slice()
-          .sort((a: Record<string, any>, b: Record<string, any>) =>
-            a.description > b.description ? 1 : -1
-          );
-      }
-      return results;
+      const filteredResults = results.filter((data) => data?.machines !== null);
+      return filteredResults;
     } catch (err) {
       throw new Error(mapError(err));
     }
@@ -112,37 +94,63 @@ export class MachineWorkCenterResolver {
     @Arg('releaseNo') releaseNo: string
   ): Promise<MachineWorkCenter[] | undefined> {
     try {
-      const sql = `
-        SELECT   wc.work_center_no AS "workCenterNo",
-                 wc.contract       AS "contract",
-                 wc.description    AS "description",
-                 department_no     AS "departmentNo",
-                 ram.machine_id    AS "machineId",
-                 ram.description   AS "machineDescription",
-                 wc.objid          AS "objId"
-        FROM     atj_work_center_v  wc,
-                 rob_apm_machine    ram
-        WHERE    ram.contract = wc.contract
-        AND      ram.work_center_no = wc.work_center_no
-        AND      wc.contract = :contract
-        AND      (wc.contract, wc.work_center_no) NOT IN
-                  (SELECT ramt.contract,
-                          rob_apm_machine_api.get_work_center_no(ramt.machine_id, ramt.contract)
-                          AS work_center_no
-                   FROM   rob_apm_maintenance ramt
-                   WHERE  ramt.contract = wc.contract
-                   AND    ramt.pr_no = :requisition_no
-                   AND    ramt.pr_line_no = :line_no
-                   AND    ramt.pr_release_no = :release_no)
-        ORDER BY wc.work_center_no
-    `;
-      const results = await getConnection().query(sql, [
-        contract,
-        requisitionNo,
-        lineNo,
-        releaseNo
-      ]);
-      return results;
+      const unavailableWC = await MachineMaintenanceView.find({
+        select: { workCenterNo: true },
+        where: {
+          contract,
+          prNo: requisitionNo,
+          prLineNo: lineNo,
+          prReleaseNo: releaseNo
+        }
+      });
+      const unavailableWCArray = unavailableWC.map((data) => data.workCenterNo);
+      let sql = '';
+      if (contract === 'AGT') {
+        sql = `
+          SELECT   wc.work_center_no AS "workCenterNo",
+                   wc.contract       AS "contract",
+                   wc.description    AS "description",
+                   wc.department_no  AS "departmentNo",
+                   ram.machine_id    AS "machineId",
+                   ram.description   AS "machineDescription",
+                   wc.objid          AS "objId"
+          FROM     work_center@ifs8agt wc,
+                   rob_apm_machine     ram
+          WHERE    ram.contract = wc.contract
+          AND      ram.work_center_no = wc.work_center_no
+          AND      wc.contract = :contract
+        `;
+      } else {
+        sql = `
+          SELECT   wc.work_center_no AS "workCenterNo",
+                   wc.contract       AS "contract",
+                   wc.description    AS "description",
+                   wc.department_no  AS "departmentNo",
+                   ram.machine_id    AS "machineId",
+                   ram.description   AS "machineDescription",
+                   wc.objid          AS "objId"
+          FROM     work_center       wc,
+                   rob_apm_machine   ram
+          WHERE    ram.contract = wc.contract
+          AND      ram.work_center_no = wc.work_center_no
+          AND      wc.contract = :contract
+        `;
+      }
+      const result = await ifs.query(sql, [contract]);
+      const excludedMachCategory = ['HD', 'HJ'];
+      let filteredResults = result.filter(
+        (data: MachineWorkCenter) =>
+          !unavailableWCArray.includes(data.workCenterNo) &&
+          !excludedMachCategory.includes(data.machineId.substring(0, 2))
+      );
+      if (filteredResults) {
+        filteredResults = filteredResults
+          .slice()
+          .sort((a: Record<string, any>, b: Record<string, any>) =>
+            a.description > b.description ? 1 : -1
+          );
+      }
+      return filteredResults;
     } catch (err) {
       throw new Error(mapError(err));
     }