+++ src/modules/manufacturing/benang-sisa-do.rv.ts
@@ -1,7 +1,8 @@
+import { ifs } from '@/config/data-sources';
 import { isAuth } from '@/middlewares/is-auth';
 import { mapError } from '@/utils/map-error';
 import { Arg, Mutation, Query, Resolver, UseMiddleware } from 'type-graphql';
-import { getConnection, In } from 'typeorm';
+import { In } from 'typeorm';
 import { BenangSisaDoInput } from './benang-sisa-do.in';
 import { BenangSisaDo } from './entities/benang-sisa-do';
 
@@ -12,7 +13,7 @@ export class BenangSisaDoResolver {
   async getBenangSisaDo(
     @Arg('contract', () => [String]) contract: string[]
   ): Promise<BenangSisaDo[] | undefined> {
-    return await BenangSisaDo.find({
+    return await BenangSisaDo.findBy({
       contract: In(contract)
     });
   }
@@ -23,7 +24,7 @@ export class BenangSisaDoResolver {
     @Arg('input') input: BenangSisaDoInput
   ): Promise<BenangSisaDo | undefined> {
     try {
-      const check = await BenangSisaDo.findOne({
+      const check = await BenangSisaDo.findOneBy({
         contract: input.contract,
         tanggal: input.tanggal
       });
@@ -34,7 +35,7 @@ export class BenangSisaDoResolver {
       where extract ( year from tanggal) = extract ( year from :p_date)
       and extract ( month from tanggal) = extract ( month from :p_date)
       and contract = :p_contract`;
-      const result = await getConnection().query(sql, [
+      const result = await ifs.query(sql, [
         input.tanggal,
         input.tanggal,
         input.contract
@@ -59,12 +60,12 @@ export class BenangSisaDoResolver {
     @Arg('input') input: BenangSisaDoInput
   ): Promise<BenangSisaDo | undefined | number> {
     try {
-      const data = await BenangSisaDo.findOne({
+      const data = await BenangSisaDo.findOneBy({
         contract: input.contract,
         tanggal: input.tanggal
       });
       if (!data) throw new Error('No data found.');
-      BenangSisaDo.merge(data, input);
+      BenangSisaDo.merge(data, { ...input });
       const results = await BenangSisaDo.save(data);
       return results;
     } catch (err) {
@@ -79,7 +80,7 @@ export class BenangSisaDoResolver {
     @Arg('tanggal') tanggal: Date
   ): Promise<BenangSisaDo> {
     try {
-      const data = await BenangSisaDo.findOne({
+      const data = await BenangSisaDo.findOneBy({
         contract,
         tanggal
       });