+++ src/modules/manufacturing/ifs-shop-ord.rv.ts
@@ -1,7 +1,8 @@
+import { ifs } from '@/config/data-sources';
 import { isAuth } from '@/middlewares/is-auth';
 import { mapError } from '@/utils/map-error';
 import { Arg, Query, Resolver, UseMiddleware } from 'type-graphql';
-import { getConnection, In } from 'typeorm';
+import { In } from 'typeorm';
 import { IfsShopOrderView } from './entities/ifs-shop-ord.vw';
 
 @Resolver(IfsShopOrderView)
@@ -12,7 +13,7 @@ export class IfsShopOrderResolver {
     @Arg('contract', () => [String]) contract: string[],
     @Arg('orderNo') orderNo: string
   ): Promise<IfsShopOrderView[] | undefined> {
-    return await IfsShopOrderView.find({ contract: In(contract), orderNo });
+    return await IfsShopOrderView.findBy({ contract: In(contract), orderNo });
   }
 
   @Query(() => IfsShopOrderView, { nullable: true })
@@ -22,7 +23,7 @@ export class IfsShopOrderResolver {
   ): Promise<Record<string, string | undefined>> {
     try {
       const sql = `SELECT SHOP_ORD_API.GET_PART_NO(:orderNo, '*', '*')  as "partNo" FROM DUAL`;
-      const result = await getConnection().query(sql, [orderNo]);
+      const result = await ifs.query(sql, [orderNo]);
       const partNo = result[0].partNo;
       return { partNo };
     } catch (err) {